Refactor luxury TIFF processor into package modules
