+++ luxury_tiff_batch_processor/io_utils.py
@@ -1,0 +1,482 @@
+"""I/O primitives and capability helpers for the luxury TIFF pipeline."""
+from __future__ import annotations
+
+import contextlib
+import dataclasses
+import logging
+import math
+import os
+import uuid
+from pathlib import Path
+from typing import Any, Dict, Iterable, Literal, Optional, Tuple, Union
+
+import numpy as np
+from PIL import Image
+
+try:  # Optional high-fidelity TIFF writer
+    import tifffile  # type: ignore
+except Exception:  # pragma: no cover - optional dependency
+    tifffile = None
+
+LOGGER = logging.getLogger("luxury_tiff_batch_processor")
+
+
+class LuxuryGradeException(RuntimeError):
+    """Raised when the processing environment cannot meet luxury standards."""
+
+
+class ProcessingCapabilities:
+    """Introspects optional dependencies to describe processing fidelity."""
+
+    _SENTINEL = object()
+
+    def __init__(self, tifffile_module: Any | None | object = _SENTINEL) -> None:
+        """Initialise capability detection."""
+
+        if tifffile_module is self._SENTINEL:
+            self._tifffile = tifffile
+        else:
+            self._tifffile = tifffile_module
+        self.bit_depth = 16 if self._supports_16_bit_output() else 8
+        self.hdr_capable = self._detect_hdr_support()
+
+    def _supports_16_bit_output(self) -> bool:
+        """Return ``True`` when a writer capable of 16-bit output is available."""
+
+        return bool(getattr(self._tifffile, "imwrite", None))
+
+    def _detect_hdr_support(self) -> bool:
+        """Best-effort check for HDR support given optional dependencies."""
+
+        if not self._supports_16_bit_output():
+            return False
+
+        supports_hdr = getattr(self._tifffile, "supports_hdr", True)
+        try:
+            return bool(supports_hdr)
+        except Exception:  # pragma: no cover - defensive fallback
+            return False
+
+    def assert_luxury_grade(self) -> None:
+        """Validate that the environment meets luxury-grade requirements."""
+
+        if self.bit_depth < 16:
+            raise LuxuryGradeException(
+                "Material Response requires 16-bit precision. "
+                "Install tifffile to unlock quantum color depth."
+            )
+        if not self.hdr_capable:
+            raise LuxuryGradeException(
+                "Luxury-grade processing requires HDR capability. "
+                "Install tifffile or ensure your environment supports HDR."
+            )
+
+
+@dataclasses.dataclass
+class ProcessingContext:
+    """Context manager staging output beside the destination."""
+
+    destination: Path
+    suffix: str = ".tmp"
+
+    def __post_init__(self) -> None:
+        self._staged_path: Optional[Path] = None
+
+    def _temp_path(self) -> Path:
+        unique = uuid.uuid4().hex
+        name = f".{self.destination.name}{self.suffix}-{unique}"
+        return self.destination.parent / name
+
+    def __enter__(self) -> Path:
+        self.destination.parent.mkdir(parents=True, exist_ok=True)
+        self._staged_path = self._temp_path()
+        return self._staged_path
+
+    def __exit__(self, exc_type, exc, tb) -> bool:
+        if self._staged_path is None:
+            return False
+
+        staged = self._staged_path
+        self._staged_path = None
+
+        if exc_type is None:
+            try:
+                os.replace(staged, self.destination)
+            except Exception:
+                with contextlib.suppress(Exception):
+                    staged.unlink()
+                raise
+        else:
+            with contextlib.suppress(FileNotFoundError):
+                staged.unlink()
+        return False
+
+
+@dataclasses.dataclass(frozen=True)
+class FloatDynamicRange:
+    """Describes how floating point image data was normalised."""
+
+    offset: np.ndarray
+    scale: np.ndarray
+    scale_recip: np.ndarray
+
+    @classmethod
+    def from_array(cls, arr: np.ndarray) -> Optional["FloatDynamicRange"]:
+        """Create a descriptor capturing per-channel offset and scale."""
+
+        if arr.size == 0:
+            return None
+
+        if arr.ndim == 2:
+            working = arr[:, :, None]
+        else:
+            working = arr
+
+        channels = working.shape[-1]
+        flattened = working.reshape(-1, channels)
+
+        offsets = np.zeros(channels, dtype=np.float32)
+        scales = np.ones(channels, dtype=np.float32)
+        saw_finite = False
+
+        for idx in range(channels):
+            channel = flattened[:, idx]
+            finite = channel[np.isfinite(channel)]
+            if finite.size == 0:
+                offsets[idx] = 0.0
+                scales[idx] = 1.0
+                continue
+
+            saw_finite = True
+            min_val = float(np.min(finite))
+            max_val = float(np.max(finite))
+            offsets[idx] = np.float32(min_val)
+            diff = float(max_val - min_val)
+            if diff <= 0.0 or not math.isfinite(diff):
+                scales[idx] = 1.0
+            else:
+                scales[idx] = np.float32(diff)
+
+        if not saw_finite:
+            return None
+
+        scale_recip = np.where(scales == 0.0, 1.0, 1.0 / scales).astype(np.float32)
+        return cls(offset=offsets, scale=scales.astype(np.float32), scale_recip=scale_recip)
+
+    @staticmethod
+    def _prepare(arr: np.ndarray) -> Tuple[np.ndarray, bool]:
+        working = np.asarray(arr, dtype=np.float32)
+        squeezed = False
+        if working.ndim == 2:
+            working = working[:, :, None]
+            squeezed = True
+        return working, squeezed
+
+    def normalise(self, arr: np.ndarray) -> np.ndarray:
+        working, squeezed = self._prepare(arr)
+        offset = self.offset.reshape((1, 1, -1))
+        scale = self.scale_recip.reshape((1, 1, -1))
+        normalised = (working - offset) * scale
+        if squeezed:
+            return normalised[:, :, 0]
+        return normalised
+
+    def denormalise(self, arr: np.ndarray) -> np.ndarray:
+        working, squeezed = self._prepare(arr)
+        offset = self.offset.reshape((1, 1, -1))
+        scale = self.scale.reshape((1, 1, -1))
+        restored = working * scale + offset
+        if squeezed:
+            return restored[:, :, 0]
+        return restored
+
+
+@dataclasses.dataclass(frozen=True)
+class ImageToFloatResult:
+    """Container for :func:`image_to_float` results with metadata."""
+
+    array: np.ndarray
+    dtype: np.dtype
+    alpha: Optional[np.ndarray]
+    base_channels: int
+    float_normalisation: Optional[FloatDynamicRange] = None
+
+    def __iter__(self):
+        yield self.array
+        yield self.dtype
+        yield self.alpha
+        yield self.base_channels
+
+
+def image_to_float(
+    image: Image.Image,
+    return_format: Literal["tuple3", "tuple4", "object"] = "object",
+) -> Union[
+    Tuple[np.ndarray, np.dtype, Optional[np.ndarray]],
+    Tuple[np.ndarray, np.dtype, Optional[np.ndarray], int],
+    ImageToFloatResult,
+]:
+    """Convert a PIL image to an RGB float32 array in the 0-1 range."""
+
+    supported_modes = {
+        "RGB",
+        "RGBA",
+        "I",
+        "I;16",
+        "I;16L",
+        "I;16B",
+        "I;16S",
+        "F",
+        "L",
+        "LA",
+    }
+    if image.mode not in supported_modes:
+        image = image.convert("RGBA" if "A" in image.getbands() else "RGB")
+
+    arr = np.array(image)
+    alpha_channel: Optional[np.ndarray] = None
+
+    if arr.ndim == 2:
+        base_channels = 1
+        color_data = arr[:, :, None]
+    else:
+        if arr.shape[2] == 4:
+            alpha_channel = arr[:, :, 3]
+            base_channels = 3
+            color_data = arr[:, :, :3]
+        elif arr.shape[2] == 2:
+            alpha_channel = arr[:, :, 1]
+            base_channels = 1
+            color_data = arr[:, :, :1]
+        else:
+            base_channels = arr.shape[2]
+            color_data = arr[:, :, :base_channels]
+
+    color_float = color_data.astype(np.float32, copy=False)
+    alpha_float = alpha_channel.astype(np.float32, copy=False) if alpha_channel is not None else None
+
+    float_norm: Optional[FloatDynamicRange] = None
+    if np.issubdtype(color_data.dtype, np.integer):
+        dtype_info = np.iinfo(color_data.dtype)
+        scale = dtype_info.max - dtype_info.min
+        if scale == 0:
+            normalised = np.zeros_like(color_float, dtype=np.float32)
+            if alpha_float is not None:
+                alpha_float = np.zeros_like(alpha_float, dtype=np.float32)
+        else:
+            offset = float(dtype_info.min)
+            inv_scale = 1.0 / float(scale)
+            normalised = (color_float - offset) * inv_scale
+            if alpha_float is not None:
+                alpha_float = (alpha_float - offset) * inv_scale
+        normalised = np.clip(normalised, 0.0, 1.0)
+        if alpha_float is not None:
+            alpha_float = np.clip(alpha_float, 0.0, 1.0)
+    else:
+        float_norm = FloatDynamicRange.from_array(color_float)
+        if float_norm is not None:
+            normalised = float_norm.normalise(color_float)
+        else:
+            normalised = color_float
+        normalised = np.clip(normalised, 0.0, 1.0)
+        if alpha_float is not None:
+            alpha_float = np.clip(alpha_float, 0.0, 1.0)
+
+    if base_channels == 1:
+        working = np.repeat(normalised, 3, axis=2)
+    else:
+        working = normalised
+
+    result = ImageToFloatResult(
+        array=np.ascontiguousarray(working, dtype=np.float32),
+        dtype=np.dtype(color_data.dtype),
+        alpha=None if alpha_float is None else np.ascontiguousarray(alpha_float, dtype=np.float32),
+        base_channels=base_channels,
+        float_normalisation=float_norm,
+    )
+
+    if return_format == "tuple3":
+        return result.array, result.dtype, result.alpha
+    if return_format == "tuple4":
+        return result.array, result.dtype, result.alpha, result.base_channels
+    allowed_formats = ("object", "tuple3", "tuple4")
+    if return_format not in allowed_formats:
+        raise ValueError(
+            f"Unsupported return_format: {return_format}. Allowed values are: {allowed_formats}"
+        )
+    return result
+
+
+def float_to_dtype_array(
+    arr: np.ndarray,
+    dtype: np.dtype,
+    alpha: Optional[np.ndarray],
+    base_channels: Optional[int] = None,
+    *,
+    float_normalisation: Optional[FloatDynamicRange] = None,
+) -> np.ndarray:
+    arr = np.clip(arr, 0.0, 1.0)
+    if arr.ndim == 2:
+        working = arr[:, :, None]
+    else:
+        working = arr
+
+    if base_channels is None:
+        base_channels = working.shape[2]
+    color = working[:, :, :base_channels]
+
+    np_dtype = np.dtype(dtype)
+    dtype_info = np.iinfo(np_dtype) if np.issubdtype(np_dtype, np.integer) else None
+    if dtype_info:
+        scale = float(dtype_info.max - dtype_info.min)
+        if scale == 0:
+            color_int = np.full_like(color, dtype_info.min, dtype=np_dtype)
+        else:
+            color_int = np.round(color * scale + dtype_info.min).astype(np_dtype)
+    else:
+        if float_normalisation is not None and np.issubdtype(np_dtype, np.floating):
+            color = float_normalisation.denormalise(color)
+        color_int = color.astype(np_dtype, copy=False)
+
+    channels: list[np.ndarray] = [color_int]
+
+    if alpha is not None:
+        alpha = np.clip(alpha, 0.0, 1.0)
+        if dtype_info:
+            scale = float(dtype_info.max - dtype_info.min)
+            if scale == 0:
+                alpha_int = np.full_like(alpha, dtype_info.min, dtype=np_dtype)
+            else:
+                alpha_int = np.round(alpha * scale + dtype_info.min).astype(np_dtype)
+        else:
+            alpha_int = alpha.astype(np_dtype, copy=False)
+        channels.append(alpha_int[:, :, None])
+
+    result = np.concatenate(channels, axis=2) if len(channels) > 1 else color_int
+    if result.shape[2] == 1:
+        result = result[:, :, 0]
+    return np.ascontiguousarray(result)
+
+
+def compression_for_tifffile(compression: str) -> Optional[str]:
+    comp = compression.lower()
+    mapping = {
+        "tiff_lzw": "lzw",
+        "lzw": "lzw",
+        "tiff_adobe_deflate": "deflate",
+        "adobe_deflate": "deflate",
+        "deflate": "deflate",
+        "tiff_zip": "deflate",
+        "zip": "deflate",
+        "tiff_jpeg": "jpeg",
+        "jpeg": "jpeg",
+        "tiff_none": None,
+        "none": None,
+        "raw": None,
+    }
+    return mapping.get(comp, comp)
+
+
+def sanitize_tiff_metadata(raw_metadata: Optional[Any]) -> Optional[Dict[int, Any]]:
+    if raw_metadata is None:
+        return None
+    safe: Dict[int, Any] = {}
+    forbidden_tags = {256, 257, 273, 279, 322, 323, 324, 325}
+    try:
+        for tag in raw_metadata:
+            if tag in forbidden_tags:
+                continue
+            safe[tag] = raw_metadata[tag]
+    except Exception:  # pragma: no cover - metadata best effort
+        LOGGER.debug("Unable to sanitise TIFF metadata", exc_info=True)
+        return None
+    return safe or None
+
+
+def save_image(
+    destination: Path,
+    arr_int: np.ndarray,
+    dtype: np.dtype,
+    metadata,
+    icc_profile: Optional[bytes],
+    compression: str,
+) -> None:
+    destination_fs = os.fspath(destination)
+    metadata = sanitize_tiff_metadata(metadata)
+    np_dtype = np.dtype(dtype)
+    dtype_info = np.iinfo(np_dtype) if np.issubdtype(np_dtype, np.integer) else None
+
+    use_tifffile = tifffile is not None and (
+        (dtype_info is not None and dtype_info.bits >= 16)
+        or np.issubdtype(np_dtype, np.floating)
+    )
+
+    array_to_write = arr_int
+    if array_to_write.ndim == 2:
+        photometric = "minisblack"
+    else:
+        photometric = "rgb" if array_to_write.shape[2] >= 3 else "minisblack"
+
+    if use_tifffile:
+        tiff_kwargs = {
+            "photometric": photometric,
+            "compression": compression_for_tifffile(compression),
+            "metadata": None,
+        }
+        if array_to_write.ndim == 3 and array_to_write.shape[2] > 3:
+            tiff_kwargs["extrasamples"] = "unassoc"
+        extratags = []
+        if icc_profile:
+            extratags.append((34675, "B", len(icc_profile), icc_profile, False))
+        if metadata:
+            try:
+                tiff_kwargs["metadata"] = {tag: metadata[tag] for tag in metadata}
+            except Exception:  # pragma: no cover - best-effort metadata copy
+                LOGGER.debug("Unable to serialise TIFF metadata", exc_info=True)
+        if extratags:
+            tiff_kwargs["extratags"] = extratags
+        tifffile.imwrite(destination_fs, array_to_write, **tiff_kwargs)
+        return
+
+    if dtype_info and dtype_info.bits == 16:
+        LOGGER.warning(
+            "Falling back to Pillow for 16-bit save; output will be 8-bit. Install 'tifffile' for full 16-bit support."
+        )
+        scale = dtype_info.max / 255.0 if dtype_info.max else 1.0
+        rgb = np.clip(array_to_write[..., :3], 0, dtype_info.max).astype(np.float32) / scale
+        rgb8 = np.clip(np.round(rgb), 0, 255).astype(np.uint8)
+        if array_to_write.ndim == 3 and array_to_write.shape[2] > 3:
+            alpha = np.clip(array_to_write[..., 3], 0, dtype_info.max).astype(np.float32) / scale
+            alpha8 = np.clip(np.round(alpha), 0, 255).astype(np.uint8)
+            array_to_write = np.concatenate([rgb8, alpha8[:, :, None]], axis=2)
+        else:
+            array_to_write = rgb8
+    elif dtype_info and dtype_info.bits < 16 and array_to_write.ndim == 3 and array_to_write.shape[2] > 3:
+        array_to_write = array_to_write.astype(np.uint8)
+
+    if array_to_write.ndim == 3 and array_to_write.shape[2] == 4:
+        mode = "RGBA"
+    elif array_to_write.ndim == 3 and array_to_write.shape[2] == 3:
+        mode = "RGB"
+    else:
+        mode = "L"
+
+    image = Image.fromarray(array_to_write, mode=mode)
+    save_kwargs = {"compression": compression}
+    if metadata is not None:
+        save_kwargs["tiffinfo"] = metadata
+    if icc_profile:
+        save_kwargs["icc_profile"] = icc_profile
+    image.save(destination_fs, format="TIFF", **save_kwargs)
+
+
+__all__ = [
+    "FloatDynamicRange",
+    "ImageToFloatResult",
+    "LuxuryGradeException",
+    "ProcessingCapabilities",
+    "ProcessingContext",
+    "float_to_dtype_array",
+    "image_to_float",
+    "save_image",
+]
